I added two classes.  Drive and Spin which uses DriveForwardForDistance and SpinInPlace.
I changed the autonomous code in Robot.java to DriveAndSpin.

I added the ROBOTWIDTH and ROBOTLENGTH parameters to RobotMap.

Finally I added a not really functional function to the Sensors class.
It is supposed to straighten out the robot's movement, but it doesn't
really work.  I am probably going to work on it after next workshop.